move simulation tests to a separate directory
